% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_editDT.R
\name{editDT_ui}
\alias{editDT_ui}
\alias{editDT_server}
\title{Editable DataTable {shiny} Module}
\usage{
editDT_ui(id)

editDT_server(
  id,
  inputTableFull,
  inputModifyPlaceID,
  colToDisplay = NULL,
  colToEdit = NULL,
  colWidth = NULL,
  colSort = NULL,
  indexStart = 0
)
}
\arguments{
\item{id}{id for module. The id should match in both the
editDT_ui and editDT_server function calls.}

\item{inputTableFull}{Main input table.Must be initialized with
\code{editDT_prepareNewData()}. Must be passed in a
\code{reactive()} statement}

\item{inputModifyPlaceID}{Modify ID. Initialized at 0. Must be passed in a
\code{reactive()} statement}

\item{colToDisplay}{Character vector of column names to display in the main
table.}

\item{colToEdit}{Character vector of column names that will be editable when
an edit button is pressed.}

\item{colWidth}{Named numeric vector for the width of displayed columns.}

\item{colSort}{Named logical vector for whether sorting is enabled on a
columns. (Currently Unimplemented)}

\item{indexStart}{Whether the index should start at 0 or 1. If the table
contains all new data (i.e. all from a form entry in a current session)
set the index to 1, else set it to 0 (the default)}

\item{input, output, session}{Internal parameters for {shiny}. These don't have
to be specified.}
}
\value{
editDT_server retruns a reactiveValues list containing
  the modified input table named 'tableFull' and the modifyID named 'modifyPlaceID'.
  These returns are used to allow the main inputs to be adapted and changed
  inside and outside the module.
}
\description{
A shiny Module to create an editable DataTable. It adds buttons
  at the top of the table for deleting selected rows, undoing, and redoing
  modifications to the table including deletes, additions, and edits,
  deletions. Additionally it adds a new row to the front of the dataset that
  holds a row specific delete and edit buttons.
}
\section{Creation notes}{
 This module was first created 2020-12-11 when
  designing the CovidDataEntry Shiny application.
}

\examples{
\dontrun{

# See the example file in the repo at /dev/run_mod_editableDataTable.R for an example of how this module works

}

}
