% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_editDT.R
\name{editDT_currentData}
\alias{editDT_currentData}
\alias{editDT_displayData}
\alias{editDT_deletedData}
\alias{editDT_prepareNewData}
\alias{editDT_rebaseModifyPoint}
\title{editDT Helpers}
\usage{
editDT_currentData(fullData, modifyPlaceID)

editDT_displayData(fullData, modifyPlaceID)

editDT_deletedData(fullData, modifyPlaceID)

editDT_prepareNewData(
  newData,
  modifyPlaceID = 0,
  fullData = NULL,
  newRecord = FALSE
)

editDT_rebaseModifyPoint(fullData, modifyPlaceID, indexStart = 0)
}
\arguments{
\item{fullData}{A data.frame/tibble previously initialized by a
\code{editDT_prepareNewData()} call and stored in a \code{reactiveValues()}
list.}

\item{modifyPlaceID}{The id for the current place in the modify timeline
being tracked using the editDT_modifyID column of fullData. This is an integer
value from 0 to infinity (i.e. zero indexed). Default 0 used for
initializing a dataset for use in the module.}

\item{newData}{A data.frame/tibble that holds a new dataset for being
initialized for or added into a \code{reactiveValues()} object for use in
the module.}

\item{newRecord}{TRUE/FALSE of whether or not the record is new. When
initializing data from your main source (i.e., a database or file) the
default is FALSE. When adding a new record to an already initialized module
dataset, set to TRUE.}

\item{indexStart}{Whether the index should start at 0 or 1. If the table
contains all new data (i.e. all from a form entry in a current session)
set the index to 1, else set it to 0 (the default)}
}
\description{
Functions that can be used inside or outside the module to add,
  modify, or filter and display the dataset that gets passed inside and
  outside the module.
}
\section{Function Specific Details}{

\describe{
  \item{editDT_currentData}{
    \itemize{
      \item Filters out the current data based on your modifyPlaceID. This
      does not filter out deleted data (see \code{editDT_displayData()})
      \item For use inside and outside the module.}}
  \item{editDT_displayData}{
    \itemize{
      \item Filters out the current data based on your modifyPlaceID and
      filters out deleted data.
      \item This should mimic the data being used to fill the module output
      DataTable.
      \item For use inside and outside the module.}}
  \item{editDT_deletedData}{
    \itemize{
      \item Filters out the current data based on your modifyPlaceID and
      returns only the currently deleted rows.
      \item This is used in the module to determine if there are any deleted
      rows.
      \item For use inside and outside the module.}}
  \item{editDT_prepareNewData}{
   \itemize{
      \item Prepares data for input into module
      by adding the columns: 'editDT_uuid', 'editDT_deleteState',
      'editDT_editState', 'editDT_newRecord', 'editDT_rowID',
      'editDT_modifyID'. Can be used to initialize a dataset for the module
      or prepare a new record to be added to an already initialized dataset
      for the module.
      \item For use outside the module.
      \item The argument newData and modifyPlaceID are used to initialize
      the input to the main 'tableFull' reactive value.
      \item When initializing a new dataset set the modifyPlaceID to 0.
      \item When using the function to prepare a new record for input into
      the table, all arguments will be used and the newRecord argument
      should be set to TRUE. The fullData argument is used to input the in
      already initialized reactive values dataset}}
  \item{editDT_rebaseModifyPoint}{
    \itemize{
      \item Determines where you are in the modify timeline.If your
      modifyPlaceID is not equal to \code{max(fullData[['editDT_modifyID']])} it
      will rebase the fullData. This involves removing all records that have
      a editDT_modifyID larger the modifyPlaceID and redefining the editDT_modifyID
      timeline to rebase to 0.
      \item WARNING: This is a destructive action and the module is not
      equipped to undo this modification.
      \item This function is called before every action that required a
      record to be appended to the fullData. This includes record adds,
      deletes, and edits.
      \item For use inside and outside the module.}}
}
}

